+++ config/database.d.ts
@@ -1,3 +1,4 @@
 import type { Knex } from "knex";
 declare const config: Knex.Config;
+export declare const initDatabase: () => Knex;
 export default config;