+++ config/database.js
@@ -1,3 +1,6 @@
+import fs from "fs";
+import knex from "knex";
+import path from "path";
 const config = {
     client: "pg",
     connection: {
@@ -20,5 +23,14 @@ const config = {
         extension: "ts",
         directory: "./database/seeders"
     }
+};
+export const initDatabase = () => {
+    const configPath = path.resolve(process.cwd(), "config/database.ts");
+    let _config;
+    if (fs.existsSync(configPath))
+        _config = require(configPath).default;
+    else
+        _config = config;
+    return knex(_config);
 };
 export default config;