+++ src/config/database.ts
@@ -1,4 +1,7 @@
 import type {Knex} from "knex";
+import fs from "fs";
+import knex from "knex";
+import path from "path";
 
 const config: Knex.Config = {
     client: "pg",
@@ -22,6 +25,17 @@ const config: Knex.Config = {
         extension: "ts",
         directory: "./database/seeders"
     }
+};
+
+export const initDatabase = (): Knex => {
+    const configPath = path.resolve(process.cwd(), "config/database.ts");
+
+    let _config: any;
+
+    if (fs.existsSync(configPath)) _config = require(configPath).default;
+    else _config = config;
+
+    return knex(_config);
 };
 
 export default config;